backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
huggingface_hub
